+++ docs/make.jl
@@ -13,6 +13,7 @@ function process_literate(names...)
 end
 
 basic_mds    = process_literate("introduction", "costfunctions")
+advanced_mds = process_literate("combined")
 
 makedocs(;
     modules=[Minuit2],
@@ -25,7 +26,8 @@ makedocs(;
     pages=[
         "Introduction" => "index.md",
         "Public API" => "api.md",
-        "Tutorials" => [ "Basic" => basic_mds ],
+        "Tutorials" => [ "Basic" => basic_mds,
+                        "Advanced" => advanced_mds],
         "Release Notes" => "release_notes.md",
     ],
     checkdocs=:exports,