+++ src/Minuit2.jl
@@ -20,6 +20,7 @@ module Minuit2
     include("wrap.jl")
     include("cost.jl")
     include("api.jl")
+    include("distributions.jl")
 
     export draw_contour, draw_mncontour, draw_profile, draw_mnprofile, visualize
     function draw_contour end